battleDisaster() function works now
Instead of removing disasters from the disaster list there is a check inside of Draw() to check the health of disaster objects. If Disaster health is above 0 the sprite is going to be drawn every frame and if the disaster health is less than 0 the Sprite for disaster will not be drawn.